// src/types/geoJS.d.ts
export type GeoJSLocation = {
  accuracy: number;
  area_code: string;
  asn: number;
  city: string;
  continent_code: string;
  country: string;
  country_code: string;
  country_code3: string;
  ip: string;
  latitude: string;
  longitude: string;
  organization: string;
  organization_name: string;
  region: string;
  timezone: string;
};